added eslint figuring out errors
