--- src/components/Layout.js
@@ -1,19 +1,0 @@
-import { Box } from '@mui/material';
-import React from 'react';
-import Header from '../components/Header';
-
-const Layout = (props) => {
-    const { children, shoppingCart } = props;
-
-    return (
-        <Box minWidth={320}>
-            <Header shoppingCart={shoppingCart} />
-            <div>
-                {children}
-            </div>
-        </Box>
-    )
-
-};
-
-export default Layout;
